requirements.txt
# =========================
# Core
# =========================
numpy>=1.26,<3
pandas>=2.1,<3
pyarrow>=14
PyYAML>=6,<7
joblib>=1.3,<2
scikit-learn>=1.3,<2
lightgbm>=4.0,<5

# =========================
# Explainability & plotting
# =========================
shap>=0.41,<1
matplotlib>=3.8,<4
scipy>=1.10,<2
numba>=0.57,<0.60

# =========================
# App (local UI)
# =========================
streamlit>=1.33

# =========================
# Pipeline / data versioning
# =========================
dvc>=3.0

# =========================
# FastAPI API (new)
# =========================
fastapi>=0.110,<1
pydantic>=2.6,<3
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<0.0.10    # for form/file endpoints if you add them
orjson>=3.10,<4                    # optional, speeds up JSON (FastAPI auto-uses if present)

# =========================
# Optional niceties
# =========================
tqdm>=4.66